fix download image modal request permission mobile
